requirements.txt
pandas==2.3.1
openpyxl==3.1.2
flask==2.3.3
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0